add a multiselect option for ForumDropdown
